+++ src/collective/behavior/talcondition/tests/test_behavior.py
@@ -7,37 +7,41 @@ from collective.behavior.talcondition.behavior import ITALCondition
 
 class TestBehavior(IntegrationTestCase):
 
-    def test_behavior(self):
-        """Test that once enabled, the behavior do the job."""
+    def setUp(self):
+        """ """
+        super(TestBehavior, self).setUp()
         # create a testitem
         login(self.portal, TEST_USER_NAME)
         self.portal.invokeFactory(id='testitem',
                                   type_name='testtype',
                                   title='Test type')
-        testitem = self.portal.testitem
-        self.assertFalse(hasattr(testitem, 'tal_condition'))
-        # adapt the testitem
-        adapted = ITALCondition(testitem)
-        # now testitem has a 'tal_condition' attribute
-        self.assertTrue(hasattr(adapted, 'tal_condition'))
+        self.testitem = self.portal.testitem
+        self.adapted = ITALCondition(self.testitem)
+
+    def test_behavior(self):
+        """Test that once enabled, the behavior do the job."""
+        self.assertFalse(hasattr(self.testitem, 'tal_condition'))
+        # adapted is ITALCondition(self.testitem)
+        # it has a 'tal_condition' attribute
+        self.assertTrue(hasattr(self.adapted, 'tal_condition'))
         # set a tal_condition and evaluate
         # this is True
-        adapted.tal_condition = u"python:context.portal_type=='testtype'"
-        self.assertTrue(adapted.evaluate())
+        self.adapted.tal_condition = u"python:context.portal_type=='testtype'"
+        self.assertTrue(self.adapted.evaluate())
         # this is False
-        adapted.tal_condition = u"python:context.portal_type=='unexisting_portal_type'"
-        self.assertFalse(adapted.evaluate())
+        self.adapted.tal_condition = u"python:context.portal_type=='unexisting_portal_type'"
+        self.assertFalse(self.adapted.evaluate())
 
     def test_wrong_condition(self):
         """In case the condition is wrong, it just returns False
            and a message is added to the Zope log."""
-        # create a testitem
-        login(self.portal, TEST_USER_NAME)
-        self.portal.invokeFactory(id='testitem',
-                                  type_name='testtype',
-                                  title='Test type')
-        testitem = self.portal.testitem
-        adapted = ITALCondition(testitem)
         # using a wrong expression does not break anything
-        adapted.tal_condition = u'python: context.some_unexisting_method()'
-        self.assertFalse(adapted.evaluate())
+        self.adapted.tal_condition = u'python: context.some_unexisting_method()'
+        self.assertFalse(self.adapted.evaluate())
+
+    def test_evaluate_extra_expr_ctx(self):
+        """The 'evaluate' method can receive a 'extra_expr_ctx' dict
+           that will extend the TAL expression context."""
+        self.adapted.tal_condition = "python: value == '122'"
+        self.assertFalse(self.adapted.evaluate())
+        self.assertTrue(self.adapted.evaluate(extra_expr_ctx={'value': '122'}))