+++ CHANGES.rst
@@ -5,9 +5,14 @@ Changelog
 0.4 (unreleased)
 ----------------
 
-- Make the tal_condition field larger (from 30 to 80) for the 
+- Make the tal_condition field larger (from 30 to 80) for the
   AT extender as well as for the DX behavior
   [gbastien]
+- Added possibility to extend TAL expression context by passing
+  an 'extra_expr_ctx' dict to utils.evaluateExpressionFor, also
+  integrated to the 'evaluate' mthode of the DX behavior
+  [gbastien]
+
 
 0.3 (2015-07-14)
 ----------------
@@ -16,6 +21,7 @@ Changelog
   [sgeulette]
 - Little optimization
   [sgeulette]
+
 
 0.2 (2015-06-18)
 ----------------
@@ -31,4 +37,3 @@ Changelog
 
 - Initial release.
   [IMIO]
-